EditDialog: Show markers for all edited items
